Added a fake rate.
The SDR setup I am using does not actually receive a strong enough
signal at the outermost edges of what it claims to receive.  When SDRs
overlap at their boundaries, this causes problems due to how
trunk-recorder chooses which recorder to use for a particular frequency.
Setting a fake rate allows me to force trunk-recorder to choose the
correct SDR.